fix: cuts 25+ seconds from load time
The heart of this change is to `class Registry` in `hordelib\nodes\comfy_controlnet_preprocessors\uniformer\mmcv\utils\registry.py`. The now stale function `Registry.infer_scope()` relied on the following call:
- filename = inspect.getmodule(inspect.stack()[2][0]).__name__
For the uninitiated, this is an extremely heavy call (>600ms) which was happening over a dozen times. This has been reworked into simply passing the calling module to `Registry(...)`.