fix: appends ` - konyconi` to civitai restored models
The author of these models deleted their account and the models were lost with it. CivitAI restored these model entries but changed the names by appending ` - konyconi` to them. This catches up with that change until we can reduce our dependence of civitai in this sense in the longer term.