feat: Allow img2img to use n_iter (#158)
* feat: Allow img2img to use n_iter
* fix: fix n_iter for inpainting and masks
* fix: Add comparison images for cn n_iter tests
* feat: add a few more debug messages post-inference
* fix: more logging at critical places
* fix: use latest sdk version for sdk inference example
* feat: use comfyui version `b4e915e74560bd2c090f9b4ed6b73b0781b7050e`
* fix: appends ` - konyconi` to civitai restored models
The author of these models deleted their account and the models were lost with it. CivitAI restored these model entries but changed the names by appending ` - konyconi` to them. This catches up with that change until we can reduce our dependence of civitai in this sense in the longer term.
* fix: add inpainting `n_iter` test images, fix typo

---------

Co-authored-by: tazlin <tazlin.on.github@gmail.com>