fix: support for upcoming comfyui system path change
Note that this the underlying change was reverted in `d14bdb18967f7413852a364747c49599de537eec` with a note that it would be soon re-implemented. This commit pins to comfyui `0c55f16c9e66eaa4915e288b34e4f848fb2d949f` which does have the impacting change (removes the `comfy` dir from the sys path by default). readding it to the path is almost entirely for the benefit of the third party nodes which need to be excised from hordelib at the earliest possible opportunity (along with the controlnet rework) and at that point the change in `config_path.py` can be reverted/removed.