fix: cascade clip vision `crop` arg mapping
An initial attempt at fixing in 1a09ddf6b9e6f199d7905ab81b59b490bfbed5dd, the dynamic mapping of inputs fore cascade remix (as is in the case with multiple input images) was not being handled appropriately. The new nodes created by horde-engine did not (re)create the `crop` argument for the CLIPVisionEncode node.

Additionally, it appears that `"crop": "center"` is the backwards-compatible change