fix: aggressively unload other models w/ large models
This should allow adjunct models, such as clip models, to be unloaded when large models are in play. This is in response to my observation that flux's clip model tends to stay in memory now on my 8gb card, but previously would smartly unload just prior to inference (and just after clip encoding). By setting a high `memory_required` I am hoping this will force this desired behavior.